separate actions file from content.js
