+++ src/constants/index.js
@@ -1,0 +1,2 @@
+export const ESC_KEY_CODE = 27
+export const JACKUP_HEIGHT = 30
