+++ src/content/actions/index.js
@@ -1,0 +1,13 @@
+import changeFixedElementToAbsolute from './changeFixedElementToAbsolute'
+import gyazocaptureWindow from './gyazocaptureWindow'
+import gyazoCaptureSelectedArea from './gyazoCaptureSelectedArea'
+import gyazoSelectElm from './gyazoSelectElm'
+import gyazoWholeCapture from './gyazoWholeCapture'
+
+export {
+  changeFixedElementToAbsolute,
+  gyazocaptureWindow,
+  gyazoCaptureSelectedArea,
+  gyazoSelectElm,
+  gyazoWholeCapture
+}
