+++ src/content/content.js
@@ -1,587 +1,26 @@
-import thenChrome from 'then-chrome'
 import browserInfo from 'bowser'
-import storage from '../libs/storageSwitcher'
+import MessageListener from '../libs/MessageListener'
 import expander from './expander'
 import notification from './notification'
-import isPressCommandKey from '../libs/isPressCommandKey'
-import changeFixedElementToAbsolute from '../libs/changeFixedElementToAbsolute'
-import restoreFixedElement from '../libs/restoreFixedElement'
-import {lockScroll, unlockScroll, packScrollBar} from '../libs/scroll'
-import getZoomAndScale from '../libs/getZoomAndScale'
+import insertMenu from './insertMenu'
+import {changeFixedElementToAbsolute} from './actions'
 
 (function () {
   if (window.__embededGyazoContentJS) {
     return
   }
   window.__embededGyazoContentJS = true
-  const ESC_KEY_CODE = 27
-  const JACKUP_HEIGHT = 30
-  const REMOVE_GYAZOMENU_EVENT = new window.Event('removeGyazoMenu')
+  const onMessageListener = new MessageListener('content')
 
   if (/gyazo\.com/.test(location.hostname)) {
     document.documentElement.setAttribute('data-extension-installed', true)
   }
 
-  chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
-    const actions = {
-      notification: () => notification(request, sender, sendResponse),
-      insertMenu: function () {
-        let gyazoMenu = document.querySelector('.gyazo-menu:not(.gyazo-notification)')
-        if (gyazoMenu) {
-          document.body.removeChild(gyazoMenu)
-          window.dispatchEvent(REMOVE_GYAZOMENU_EVENT)
-          return true
-        }
-        let hideMenu = function () {
-          if (document.body.contains(gyazoMenu)) {
-            document.body.removeChild(gyazoMenu)
-          }
-          window.dispatchEvent(REMOVE_GYAZOMENU_EVENT)
-        }
-        gyazoMenu = document.createElement('div')
-        gyazoMenu.className = 'gyazo-menu gyazo-menu-element'
+  onMessageListener.add('notification', notification)
+  onMessageListener.add('insertMenu', insertMenu)
+  onMessageListener.add('changeFixedElementToAbsolute', changeFixedElementToAbsolute)
 
-        let createButton = function (loadSvgName, text, shortcutKey) {
-          let btn = document.createElement('div')
-          btn.className = 'gyazo-big-button gyazo-button gyazo-menu-element'
-
-          if (shortcutKey) {
-            btn.setAttribute('title', 'Press: ' + shortcutKey)
-          }
-
-          let iconElm = document.createElement('div')
-          iconElm.classList.add('gyazo-button-icon')
-          try {
-            // Edge cannot fetch to ms-edge-extension:
-            window.fetch(chrome.runtime.getURL(`imgs/${loadSvgName}.svg`))
-              .then((res) => res.text())
-              .then((text) => { iconElm.innerHTML = text })
-          } catch (e) {
-            const svgUrl = chrome.runtime.getURL(`imgs/${loadSvgName}.svg`)
-            iconElm.innerHTML = `<img src='${svgUrl}' />`
-          }
-
-          let textElm = document.createElement('div')
-          textElm.className = 'gyazo-button-text'
-          textElm.textContent = text
-
-          btn.appendChild(iconElm)
-          btn.appendChild(textElm)
-
-          return btn
-        }
-
-        let selectElementBtn = createButton('selection', chrome.i18n.getMessage('selectElement'), 'E')
-        let selectAreaBtn = createButton('crop', chrome.i18n.getMessage('selectArea'), 'S')
-        let windowCaptureBtn = createButton('window', chrome.i18n.getMessage('captureWindow'), 'P')
-        let wholeCaptureBtn = createButton('window-scroll', chrome.i18n.getMessage('topToBottom'), 'W')
-        let myImageBtn = createButton('grid', chrome.i18n.getMessage('myImage'))
-        myImageBtn.classList.add('gyazo-menu-myimage')
-        let closeBtn = document.createElement('div')
-        closeBtn.className = 'gyazo-close-button gyazo-menu-element'
-        const closeBtnIcon = document.createElement('div')
-        closeBtnIcon.className = 'gyazo-menu-element gyazo-icon gyazo-icon-cross'
-        closeBtn.appendChild(closeBtnIcon)
-        try {
-          window.fetch(chrome.runtime.getURL('imgs/cross.svg'))
-            .then((res) => res.text())
-            .then((text) => { closeBtnIcon.innerHTML = text })
-        } catch (e) {
-          closeBtnIcon.innerHTML = `<img src='${chrome.runtime.getURL('imgs/cross.svg')}' />`
-        }
-        closeBtn.setAttribute('title', 'Press: Escape')
-
-        window.addEventListener('contextmenu', function (event) {
-          hideMenu()
-        })
-        document.body.appendChild(gyazoMenu)
-        gyazoMenu.appendChild(selectElementBtn)
-        gyazoMenu.appendChild(selectAreaBtn)
-        gyazoMenu.appendChild(windowCaptureBtn)
-        gyazoMenu.appendChild(wholeCaptureBtn)
-        gyazoMenu.appendChild(myImageBtn)
-        gyazoMenu.appendChild(closeBtn)
-
-        let hotKey = function (event) {
-          window.removeEventListener('keydown', hotKey)
-          if (event.keyCode === ESC_KEY_CODE) {
-            hideMenu()
-          }
-          switch (String.fromCharCode(event.keyCode)) {
-            case 'E':
-              selectElementBtn.click()
-              break
-            case 'S':
-              selectAreaBtn.click()
-              break
-            case 'P':
-              windowCaptureBtn.click()
-              break
-            case 'W':
-              wholeCaptureBtn.click()
-              break
-          }
-        }
-        window.addEventListener('keydown', hotKey)
-        try {
-          // Provide access to chrome.storage at content script https://bugzilla.mozilla.org/show_bug.cgi?id=1197346
-          storage.get({behavior: 'element'}, function (item) {
-            if (item.behavior === 'element') {
-              // Default behavior is select element
-              selectElementBtn.classList.add('gyazo-button-active')
-              window.requestAnimationFrame(actions.gyazoSelectElm)
-            } else if (item.behavior === 'area') {
-              // Default behavior is select area
-              selectAreaBtn.classList.add('gyazo-button-active')
-              actions.gyazoCaptureSelectedArea()
-            }
-          })
-        } catch (e) {
-          selectAreaBtn.classList.add('gyazo-button-active')
-          actions.gyazoCaptureSelectedArea()
-        }
-        selectAreaBtn.addEventListener('click', function () {
-          hideMenu()
-          window.requestAnimationFrame(function () {
-            actions.gyazoCaptureSelectedArea()
-          })
-        })
-        selectElementBtn.addEventListener('click', function () {
-          hideMenu()
-          window.requestAnimationFrame(function () {
-            actions.gyazoSelectElm()
-          })
-        })
-        windowCaptureBtn.addEventListener('click', function () {
-          hideMenu()
-          window.requestAnimationFrame(function () {
-            actions.gyazocaptureWindow()
-          })
-        })
-        wholeCaptureBtn.addEventListener('click', function () {
-          hideMenu()
-          window.requestAnimationFrame(function () {
-            actions.gyazoWholeCapture()
-          })
-        })
-        closeBtn.addEventListener('click', function () {
-          hideMenu()
-        })
-        myImageBtn.addEventListener('click', function () {
-          hideMenu()
-          window.open('https://gyazo.com/')
-        })
-      },
-      changeFixedElementToAbsolute: function () {
-        changeFixedElementToAbsolute()
-        const waitScroll = function () {
-          if (Math.abs(window.scrollX - request.scrollTo.x) < 1 &&
-              Math.abs(window.scrollY - request.scrollTo.y) < 1) {
-            window.requestAnimationFrame(sendResponse)
-          } else {
-            window.requestAnimationFrame(waitScroll)
-          }
-        }
-        window.requestAnimationFrame(waitScroll)
-      },
-      gyazocaptureWindow: function () {
-        const overflow = lockScroll()
-        const data = {}
-        const scaleObj = getZoomAndScale()
-        data.w = window.innerWidth
-        data.h = window.innerHeight
-        data.x = window.scrollX
-        data.y = window.scrollY
-        data.t = document.title
-        data.u = location.href
-        data.s = scaleObj.scale
-        data.z = scaleObj.zoom
-        data.positionX = window.scrollX
-        data.positionY = window.scrollY
-        data.defaultPositon = window.scrollY
-        data.innerHeight = window.innerHeight
-        window.requestAnimationFrame(() => {
-          thenChrome.runtime.sendMessage(chrome.runtime.id, {
-            target: 'main',
-            action: 'gyazoCaptureWithSize',
-            data: data,
-            tab: request.tab
-          })
-          .then(() => {
-            unlockScroll(overflow)
-          })
-        })
-      },
-      gyazoSelectElm: function () {
-        if (document.querySelector('.gyazo-crop-select-element')) {
-          return false
-        }
-        const MARGIN = 3
-        document.body.classList.add('gyazo-select-element-mode')
-        const jackup = document.createElement('div')
-        jackup.classList.add('gyazo-jackup-element')
-        document.body.appendChild(jackup)
-        const layer = document.createElement('div')
-        layer.className = 'gyazo-crop-select-element'
-        document.body.appendChild(layer)
-        layer.style.background = 'rgba(9, 132, 222, 0.35)'
-        layer.style.margin = '0px'
-        layer.style.border = '1px solid rgb(9, 132, 222)'
-        layer.style.position = 'fixed'
-        layer.style.pointerEvents = 'none'
-        layer.style.zIndex = 2147483646 // Maximun number of 32bit Int - 1
-        const allElms = Array.from(document.body.querySelectorAll('*'))
-          .filter(function (item) {
-          return !item.classList.contains('gyazo-crop-select-element') &&
-                 !item.classList.contains('gyazo-menu-element')
-          })
-        allElms.forEach(function (item) {
-          item.classList.add('gyazo-select-element-cursor-overwrite')
-        })
-        const moveLayer = function (event) {
-          const item = event.target
-          event.stopPropagation()
-          if (item.tagName === 'IMG') {
-            layer.setAttribute('data-img-url', item.src)
-          } else {
-            layer.setAttribute('data-img-url', '')
-          }
-          const rect = item.getBoundingClientRect()
-          layer.style.width = rect.width + 'px'
-          layer.style.height = rect.height + 'px'
-          layer.style.left = rect.left + 'px'
-          layer.style.top = rect.top + 'px'
-        }
-        let hasMargin = false
-        const takeMargin = function () {
-          if (hasMargin) return
-          hasMargin = true
-          layer.style.width = parseInt(window.getComputedStyle(layer).width, 10) + MARGIN * 2 + 'px'
-          layer.style.height = parseInt(window.getComputedStyle(layer).height, 10) + MARGIN * 2 + 'px'
-          layer.style.left = parseInt(window.getComputedStyle(layer).left, 10) - MARGIN + 'px'
-          layer.style.top = parseInt(window.getComputedStyle(layer).top, 10) - MARGIN + 'px'
-        }
-        const keydownHandler = function (event) {
-          if (event.keyCode === ESC_KEY_CODE) {
-            cancel()
-          } else if (isPressCommandKey(event)) {
-            takeMargin()
-          }
-        }
-        const keyUpHandler = function (event) {
-          if (isPressCommandKey(event)) {
-            hasMargin = false
-            layer.style.width = parseInt(window.getComputedStyle(layer).width, 10) - MARGIN * 2 + 'px'
-            layer.style.height = parseInt(window.getComputedStyle(layer).height, 10) - MARGIN * 2 + 'px'
-            layer.style.left = parseInt(window.getComputedStyle(layer).left, 10) + MARGIN + 'px'
-            layer.style.top = parseInt(window.getComputedStyle(layer).top, 10) + MARGIN + 'px'
-          }
-        }
-        const clickElement = function (event) {
-          event.stopPropagation()
-          event.preventDefault()
-          document.body.classList.remove('gyazo-select-element-mode')
-          allElms.forEach(function (item) {
-            if (item.classList.contains('gyazo-select-element-cursor-overwrite')) {
-              item.classList.remove('gyazo-select-element-cursor-overwrite')
-            }
-            item.removeEventListener('mouseover', moveLayer)
-            item.removeEventListener('click', clickElement)
-          })
-          const data = {}
-          const scaleObj = getZoomAndScale()
-
-          // Sanitize gyazo desc for ivy-search
-          Array.from(document.querySelectorAll('*')).forEach(function (elm) {
-            if (window.getComputedStyle(elm).display === 'none' || window.getComputedStyle(elm).visibility === 'hidden') {
-              elm.classList.add('gyazo-hidden')
-            }
-          })
-          const dupTarget = event.target.cloneNode(true)
-          Array.from(dupTarget.querySelectorAll('*')).forEach(function (elm) {
-            switch (elm.tagName) {
-              case 'SCRIPT':
-              case 'STYLE':
-                return elm.remove()
-            }
-            if (elm.classList.contains('gyazo-hidden')) {
-              elm.remove()
-            }
-          })
-          Array.from(document.getElementsByClassName('gyazo-hidden')).forEach(function (elm) {
-            elm.classList.remove('gyazo-hidden')
-          })
-
-          data.w = parseFloat(layer.style.width)
-          data.h = parseFloat(layer.style.height)
-          data.x = window.scrollX + layer.offsetLeft
-          data.y = window.scrollY + layer.offsetTop
-          data.t = document.title
-          data.u = location.href
-          data.s = scaleObj.scale
-          data.z = scaleObj.zoom
-          data.positionX = window.scrollX
-          data.positionY = window.scrollY
-          data.innerHeight = window.innerHeight
-          data.desc = dupTarget.textContent
-          if (document.body.contains(layer)) {
-            document.body.removeChild(layer)
-          }
-          if (document.querySelector('.gyazo-menu')) {
-            document.body.removeChild(document.querySelector('.gyazo-menu'))
-          }
-          jackup.style.height = (window.innerHeight + JACKUP_HEIGHT) + 'px'
-          window.removeEventListener('contextmenu', cancel)
-          window.removeEventListener('keydown', keydownHandler)
-          document.removeEventListener('keyup', keyUpHandler)
-          if (layer.offsetTop >= 0 && layer.offsetTop + layer.offsetHeight <= window.innerHeight) {
-            // Only when required scroll
-            changeFixedElementToAbsolute()
-          }
-          if (layer.getAttribute('data-img-url')) {
-            restoreFixedElement()
-            return chrome.runtime.sendMessage(chrome.runtime.id, {
-              target: 'main',
-              action: 'gyazoSendRawImage',
-              data: {srcUrl: layer.getAttribute('data-img-url')},
-              tab: request.tab
-            }, function () {})
-          }
-          let overflow = {}
-          if (data.y + data.h > data.innerHeight + data.positionY) {
-            overflow = lockScroll()
-            packScrollBar(overflow)
-          }
-          const finish = function () {
-            if (document.getElementsByClassName('gyazo-crop-select-element').length > 0) {
-              return window.requestAnimationFrame(finish)
-            }
-            window.requestAnimationFrame(function () {
-              thenChrome.runtime.sendMessage(chrome.runtime.id, {
-                target: 'main',
-                action: 'gyazoCaptureWithSize',
-                data: data,
-                tab: request.tab
-              })
-              .then(() => {
-                restoreFixedElement()
-                document.body.removeChild(jackup)
-                unlockScroll(overflow)
-              })
-            })
-          }
-          window.requestAnimationFrame(finish)
-        }
-        const cancel = function () {
-          if (document.body.contains(jackup)) {
-            document.body.removeChild(jackup)
-          }
-          if (document.body.contains(layer)) {
-            document.body.removeChild(layer)
-          }
-          document.body.classList.remove('gyazo-select-element-mode')
-          window.removeEventListener('contextmenu', cancel)
-          document.removeEventListener('keydown', keydownHandler)
-          document.removeEventListener('keyup', keyUpHandler)
-          Array.from(document.querySelectorAll('.gyazo-select-element-cursor-overwrite'))
-            .forEach(function (item) {
-              item.classList.remove('gyazo-select-element-cursor-overwrite')
-              item.removeEventListener('mouseover', moveLayer)
-              item.removeEventListener('click', clickElement)
-            })
-          restoreFixedElement()
-        }
-        let removedGyazoMenu = function () {
-          window.removeEventListener('removeGyazoMenu', removedGyazoMenu)
-          cancel()
-        }
-        window.addEventListener('removeGyazoMenu', removedGyazoMenu)
-        window.addEventListener('contextmenu', cancel)
-        document.addEventListener('keydown', keydownHandler)
-        document.addEventListener('keyup', keyUpHandler)
-        window.requestAnimationFrame(function () {
-          allElms.forEach(function (item) {
-            item.addEventListener('mouseover', moveLayer)
-            item.addEventListener('click', clickElement)
-          })
-        })
-      },
-      gyazoCaptureSelectedArea: function () {
-        if (document.querySelector('.gyazo-jackup-element')) {
-          return false
-        }
-        let startX, startY
-        let data = {}
-        const tempUserSelect = document.body.style.webkitUserSelect
-        const layer = document.createElement('div')
-        const jackup = document.createElement('div')
-        jackup.classList.add('gyazo-jackup-element')
-        document.body.appendChild(jackup)
-        const pageHeight = Math.max(document.body.clientHeight, document.body.offsetHeight, document.body.scrollHeight)
-        layer.style.position = 'absolute'
-        layer.style.left = document.body.clientLeft + 'px'
-        layer.style.top = document.body.clientTop + 'px'
-        layer.style.width = Math.max(
-          document.body.clientWidth, document.body.offsetWidth, document.body.scrollWidth,
-          document.documentElement.clientWidth, document.documentElement.offsetWidth, document.documentElement.scrollWidth
-        ) + 'px'
-        layer.style.height = pageHeight + 'px'
-        layer.style.zIndex = 2147483646 // Maximun number of 32bit Int - 1
-        layer.style.cursor = 'crosshair'
-        layer.className = 'gyazo-select-layer'
-        document.body.style.webkitUserSelect = 'none'
-        const selectionElm = document.createElement('div')
-        layer.appendChild(selectionElm)
-        document.body.appendChild(layer)
-        selectionElm.styleUpdate = function (styles) {
-          Object.keys(styles).forEach(function (key) {
-            selectionElm.style[key] = styles[key]
-          })
-        }
-        selectionElm.styleUpdate({
-          background: 'rgba(92, 92, 92, 0.3)',
-          position: 'absolute'
-        })
-        const cancelGyazo = function () {
-          document.body.removeChild(layer)
-          document.body.removeChild(jackup)
-          document.body.style.webkitUserSelect = tempUserSelect
-          document.removeEventListener('keydown', keydownHandler)
-          window.removeEventListener('contextmenu', cancelGyazo)
-          restoreFixedElement()
-          if (document.querySelector('.gyazo-menu')) {
-            document.body.removeChild(document.querySelector('.gyazo-menu'))
-          }
-        }
-        let removedGyazoMenu = function () {
-          cancelGyazo()
-          window.removeEventListener('removeGyazoMenu', removedGyazoMenu)
-        }
-        window.addEventListener('removeGyazoMenu', removedGyazoMenu)
-        const keydownHandler = function (event) {
-          if (event.keyCode === ESC_KEY_CODE) {
-            //  If press Esc Key, cancel it
-            cancelGyazo()
-          }
-        }
-        const mousedownHandler = function (e) {
-          let gyazoMenu = document.querySelector('.gyazo-menu')
-          if (gyazoMenu) {
-            document.body.removeChild(gyazoMenu)
-          }
-          startX = e.pageX
-          startY = e.pageY
-          selectionElm.styleUpdate({
-            border: '1px solid rgba(255, 255, 255, 0.8)',
-            left: startX + 'px',
-            top: startY + 'px'
-          })
-          layer.removeEventListener('mousedown', mousedownHandler)
-          layer.addEventListener('mousemove', mousemoveHandler)
-          layer.addEventListener('mouseup', mouseupHandler)
-        }
-        const mousemoveHandler = function (e) {
-          selectionElm.styleUpdate({
-            width: (Math.abs(e.pageX - startX) - 1) + 'px',
-            height: (Math.abs(e.pageY - startY) - 1) + 'px',
-            left: Math.min(e.pageX, startX) + 'px',
-            top: Math.min(e.pageY, startY) + 'px'
-          })
-        }
-        const mouseupHandler = function (e) {
-          document.body.style.webkitUserSelect = tempUserSelect
-          document.removeEventListener('keydown', keydownHandler)
-          window.addEventListener('contextmenu', function (event) {
-            cancelGyazo()
-            event.preventDefault()
-          })
-          const scaleObj = getZoomAndScale()
-          const rect = selectionElm.getBoundingClientRect()
-          data.w = rect.width
-          data.h = rect.height
-          if (data.h <= 3 || data.w <= 3) {
-            cancelGyazo()
-            return false
-          }
-          data.x = rect.left + window.scrollX
-          data.y = rect.top + window.scrollY
-          data.t = document.title
-          data.u = location.href
-          data.s = scaleObj.scale
-          data.z = scaleObj.zoom
-          data.positionX = window.scrollX
-          data.positionY = window.scrollY
-          data.innerHeight = window.innerHeight
-          document.body.removeChild(layer)
-          if (document.querySelector('.gyazo-menu')) {
-            document.body.removeChild(document.querySelector('.gyazo-menu'))
-          }
-          let overflow = {}
-          if (data.h > data.innerHeight) {
-            overflow = lockScroll()
-            packScrollBar(overflow)
-          }
-          jackup.style.height = (window.innerHeight + JACKUP_HEIGHT) + 'px'
-          // wait for rewrite by removeChild
-          let finish = function () {
-            if (document.getElementsByClassName('gyazo-select-layer').length > 0) {
-              return window.requestAnimationFrame(finish)
-            }
-            window.setTimeout(function () {
-              chrome.runtime.sendMessage(chrome.runtime.id, {
-                target: 'main',
-                action: 'gyazoCaptureWithSize',
-                data: data,
-                tab: request.tab
-              }, function () {
-                document.body.removeChild(jackup)
-                unlockScroll(overflow)
-                restoreFixedElement()
-              })
-            }, 100)
-          }
-          window.requestAnimationFrame(finish)
-        }
-        layer.addEventListener('mousedown', mousedownHandler)
-        document.addEventListener('keydown', keydownHandler)
-        window.addEventListener('contextmenu', cancelGyazo)
-      },
-      gyazoWholeCapture: function () {
-        const overflow = lockScroll()
-        const data = {}
-        const scaleObj = getZoomAndScale()
-        data.w = window.innerWidth
-        data.h = Math.max(document.body.clientHeight, document.body.offsetHeight, document.body.scrollHeight)
-        data.x = 0
-        data.y = 0
-        data.t = document.title
-        data.u = location.href
-        data.s = scaleObj.scale
-        data.z = scaleObj.zoom
-        data.positionX = window.scrollX
-        data.positionY = window.scrollY
-        data.innerHeight = window.innerHeight
-        const jackup = document.createElement('div')
-        jackup.classList.add('gyazo-jackup-element')
-        document.body.appendChild(jackup)
-        jackup.style.height = (data.h + 30) + 'px'
-        thenChrome.runtime.sendMessage(chrome.runtime.id, {
-          target: 'main',
-          action: 'gyazoCaptureWithSize',
-          data: data,
-          tab: request.tab
-        })
-        .then(() => {
-          document.body.removeChild(jackup)
-          unlockScroll(overflow)
-        })
-      }
-    }
-    if (request.action in actions) {
-      actions[request.action]()
-    }
-    return true
-  })
+  chrome.runtime.onMessage.addListener(onMessageListener.listen.bind(onMessageListener))
   // XXX: Firefox can't embed moz-extension:// file in content
   if (!browserInfo.firefox) expander()
 })()
