+++ src/libs/restoreFixedElement.js
@@ -1,4 +1,4 @@
-() => {
+export default () => {
   const fixedElms = document.getElementsByClassName('gyazo-whole-capture-onetime-absolute')
   Array.from(fixedElms).forEach((item) => {
     item.classList.remove('gyazo-whole-capture-onetime-absolute')
